chore: check settings permission

+++ src/main/java/com/grash/controller/ChecklistController.java
@@ -7,6 +7,7 @@ import com.grash.exception.CustomException;
 import com.grash.model.Checklist;
 import com.grash.model.CompanySettings;
 import com.grash.model.OwnUser;
+import com.grash.model.enums.PermissionEntity;
 import com.grash.model.enums.RoleType;
 import com.grash.service.ChecklistService;
 import com.grash.service.UserService;
@@ -72,7 +73,7 @@ public class ChecklistController {
             @ApiResponse(code = 403, message = "Access denied")})
     public Checklist create(@ApiParam("Checklist") @Valid @RequestBody ChecklistPostDTO checklistReq, HttpServletRequest req) {
         OwnUser user = userService.whoami(req);
-        if (checklistService.canCreate(user, checklistReq)) {
+        if (checklistService.canCreate(user, checklistReq) && user.getRole().getViewPermissions().contains(PermissionEntity.SETTINGS)) {
             return checklistService.createPost(checklistReq, user.getCompany());
         } else throw new CustomException("Access denied", HttpStatus.FORBIDDEN);
     }
@@ -90,7 +91,7 @@ public class ChecklistController {
 
         if (optionalChecklist.isPresent()) {
             Checklist savedChecklist = optionalChecklist.get();
-            if (checklistService.hasAccess(user, savedChecklist) && checklistService.canPatch(user, checklist)) {
+            if (checklistService.hasAccess(user, savedChecklist) && checklistService.canPatch(user, checklist) && user.getRole().getViewPermissions().contains(PermissionEntity.SETTINGS)) {
                 return checklistService.update(id, checklist, user.getCompany());
             } else throw new CustomException("Forbidden", HttpStatus.FORBIDDEN);
         } else throw new CustomException("Checklist not found", HttpStatus.NOT_FOUND);
@@ -108,7 +109,7 @@ public class ChecklistController {
         Optional<Checklist> optionalChecklist = checklistService.findById(id);
         if (optionalChecklist.isPresent()) {
             Checklist savedChecklist = optionalChecklist.get();
-            if (checklistService.hasAccess(user, savedChecklist)) {
+            if (checklistService.hasAccess(user, savedChecklist) && user.getRole().getViewPermissions().contains(PermissionEntity.SETTINGS)) {
                 checklistService.delete(id);
                 return new ResponseEntity<>(new SuccessResponse(true, "Deleted successfully"),
                         HttpStatus.OK);
